good automation, pre-clean up. Working
