--- examples/sine_approx/memory_out.csv
@@ -1,53 +1,0 @@
-0,32768
-4,0
-100,0
-104,32428
-108,8107
-112,30412
-116,15710
-120,26504
-124,22336
-128,20950
-132,27573
-136,14092
-140,31096
-144,6358
-148,32686
-152,-1771
-156,32243
-160,-9790
-164,29796
-168,-17200
-172,25496
-176,-23541
-180,19611
-184,-28418
-188,12506
-192,-31528
-196,4624
-200,-32678
-204,-3545
-208,-31797
-212,-11494
-216,-28938
-220,-18729
-224,-24280
-228,-24799
-232,-18112
-236,-29327
-240,-10819
-244,-32032
-248,-2852
-252,-32745
-256,5291
-260,-31422
-264,13106
-268,-28146
-272,20106
-276,-23119
-280,25855
-284,-16655
-288,29998
-292,-9156
-296,32275
-10000,27573
