+++ sinusoidal/sine_approx/instructions.csv
@@ -1,0 +1,185 @@
+0
+LWD R0,LWD R0,NOP,NOP
+NOP,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+1
+"SSUB R0, R0, RCR",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+2
+"SRT R1, R0, 10",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+3
+"LAND R0, R0, 1023",NOP,NOP,NOP
+"SMUL R0, RCT, 44",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+4
+NOP,NOP,NOP,NOP
+"SADD R1, R0, 140",NOP,NOP,NOP
+"SADD R0, RCT, 100",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+5
+NOP,NOP,NOP,NOP
+"LWI R1, R1",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+6
+NOP,NOP,NOP,NOP
+"SMUL R1, R1, RCT",NOP,NOP,NOP
+"SADD ROUT, R0, 36",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+7
+NOP,NOP,NOP,NOP
+"SRA R1, R1, 10",NOP,NOP,NOP
+"LWI R1, ROUT",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+8
+NOP,NOP,NOP,NOP
+"SADD R1, R1, RCB",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+9
+NOP,NOP,NOP,NOP
+"SMUL R1, R1, RCT",NOP,NOP,NOP
+"SADD ROUT, R0, 32",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+10
+NOP,NOP,NOP,NOP
+"SRA R1, R1, 10",NOP,NOP,NOP
+"LWI R1, ROUT",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+11
+NOP,NOP,NOP,NOP
+"SADD R1, R1, RCB",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+12
+NOP,NOP,NOP,NOP
+"SMUL R1, R1, RCT",NOP,NOP,NOP
+"SADD ROUT, R0, 28",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+13
+NOP,NOP,NOP,NOP
+"SRA R1, R1, 10",NOP,NOP,NOP
+"LWI R1, ROUT",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+14
+NOP,NOP,NOP,NOP
+"SADD R1, R1, RCB",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+15
+NOP,NOP,NOP,NOP
+"SMUL R1, R1, RCT",NOP,NOP,NOP
+"SADD ROUT, R0, 24",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+16
+NOP,NOP,NOP,NOP
+"SRA R1, R1, 10",NOP,NOP,NOP
+"LWI R1, ROUT",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+17
+NOP,NOP,NOP,NOP
+"SADD R1, R1, RCB",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+18
+NOP,NOP,NOP,NOP
+"SMUL R1, R1, RCT",NOP,NOP,NOP
+"SADD ROUT, R0, 20",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+19
+NOP,NOP,NOP,NOP
+"SRA R1, R1, 10",NOP,NOP,NOP
+"LWI R1, ROUT",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+20
+NOP,NOP,NOP,NOP
+"SADD R1, R1, RCB",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+21
+NOP,NOP,NOP,NOP
+"SMUL R1, R1, RCT",NOP,NOP,NOP
+"SADD ROUT, R0, 16",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+22
+NOP,NOP,NOP,NOP
+"SRA R1, R1, 10",NOP,NOP,NOP
+"LWI R1, ROUT",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+23
+NOP,NOP,NOP,NOP
+"SADD R1, R1, RCB",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+24
+NOP,NOP,NOP,NOP
+"SMUL R1, R1, RCT",NOP,NOP,NOP
+"SADD ROUT, R0, 12",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+25
+NOP,NOP,NOP,NOP
+"SRA R1, R1, 10",NOP,NOP,NOP
+"LWI R1, ROUT",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+26
+NOP,NOP,NOP,NOP
+"SADD R1, R1, RCB",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+27
+NOP,NOP,NOP,NOP
+"SMUL R1, R1, RCT",NOP,NOP,NOP
+"SADD ROUT, R0, 8",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+28
+NOP,NOP,NOP,NOP
+"SRA R1, R1, 10",NOP,NOP,NOP
+"LWI R1, ROUT",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+29
+NOP,NOP,NOP,NOP
+"SADD R1, R1, RCB",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+30
+NOP,NOP,NOP,NOP
+"SMUL R1, R1, RCT",NOP,NOP,NOP
+"SADD ROUT, R0, 4",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+31
+NOP,NOP,NOP,NOP
+"SRA R1, R1, 10",NOP,NOP,NOP
+"LWI R1, ROUT",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+32
+NOP,NOP,NOP,NOP
+"SADD R1, R1, RCB",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+33
+NOP,NOP,NOP,NOP
+"SMUL R1, R1, RCT",NOP,NOP,NOP
+"SADD ROUT, R0, 0",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+34
+NOP,NOP,NOP,NOP
+"SRA R1, R1, 10",NOP,NOP,NOP
+"LWI R1, ROUT",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+35
+NOP,NOP,NOP,NOP
+"SADD R1, R1, RCB",NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+36
+NOP,NOP,NOP,EXIT
+SWD R1,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
+NOP,NOP,NOP,NOP
