+++ sinusoidal/sine_approx/memory_out.csv
@@ -1,0 +1,674 @@
+0,10000
+4,0
+100,0
+104,1024
+108,0
+112,-2
+116,0
+120,0
+124,0
+128,0
+132,0
+136,0
+140,0
+144,1022
+148,1019
+152,-5
+156,-2
+160,0
+164,0
+168,0
+172,0
+176,0
+180,0
+184,0
+188,2034
+192,1003
+196,-11
+200,-2
+204,0
+208,0
+212,0
+216,0
+220,0
+224,0
+228,0
+232,3024
+236,976
+240,-16
+244,-2
+248,0
+252,0
+256,0
+260,0
+264,0
+268,0
+272,0
+276,3982
+280,939
+284,-21
+288,-2
+292,0
+296,0
+300,0
+304,0
+308,0
+312,0
+316,0
+320,4899
+324,893
+328,-26
+332,-2
+336,0
+340,0
+344,0
+348,0
+352,0
+356,0
+360,0
+364,5765
+368,837
+372,-30
+376,-1
+380,0
+384,0
+388,0
+392,0
+396,0
+400,0
+404,0
+408,6570
+412,772
+416,-34
+420,-1
+424,0
+428,0
+432,0
+436,0
+440,0
+444,0
+448,0
+452,7306
+456,699
+460,-38
+464,-1
+468,0
+472,0
+476,0
+480,0
+484,0
+488,0
+492,0
+496,7966
+500,619
+504,-42
+508,-1
+512,0
+516,0
+520,0
+524,0
+528,0
+532,0
+536,0
+540,8542
+544,532
+548,-45
+552,-1
+556,0
+560,0
+564,0
+568,0
+572,0
+576,0
+580,0
+584,9029
+588,440
+592,-47
+596,-1
+600,0
+604,0
+608,0
+612,0
+616,0
+620,0
+624,0
+628,9421
+632,343
+636,-49
+640,-1
+644,0
+648,0
+652,0
+656,0
+660,0
+664,0
+668,0
+672,9714
+676,243
+680,-51
+684,0
+688,0
+692,0
+696,0
+700,0
+704,0
+708,0
+712,0
+716,9906
+720,140
+724,-52
+728,0
+732,0
+736,0
+740,0
+744,0
+748,0
+752,0
+756,0
+760,9994
+764,36
+768,-52
+772,0
+776,0
+780,0
+784,0
+788,0
+792,0
+796,0
+800,0
+804,9977
+808,-69
+812,-52
+816,0
+820,0
+824,0
+828,0
+832,0
+836,0
+840,0
+844,0
+848,9856
+852,-173
+856,-52
+860,0
+864,0
+868,0
+872,0
+876,0
+880,0
+884,0
+888,0
+892,9631
+896,-276
+900,-50
+904,0
+908,0
+912,0
+916,0
+920,0
+924,0
+928,0
+932,0
+936,9306
+940,-375
+944,-49
+948,1
+952,0
+956,0
+960,0
+964,0
+968,0
+972,0
+976,0
+980,8883
+984,-470
+988,-47
+992,1
+996,0
+1000,0
+1004,0
+1008,0
+1012,0
+1016,0
+1020,0
+1024,8367
+1028,-561
+1032,-44
+1036,1
+1040,0
+1044,0
+1048,0
+1052,0
+1056,0
+1060,0
+1064,0
+1068,7763
+1072,-645
+1076,-41
+1080,1
+1084,0
+1088,0
+1092,0
+1096,0
+1100,0
+1104,0
+1108,0
+1112,7078
+1116,-723
+1120,-37
+1124,1
+1128,0
+1132,0
+1136,0
+1140,0
+1144,0
+1148,0
+1152,0
+1156,6319
+1160,-794
+1164,-33
+1168,1
+1172,0
+1176,0
+1180,0
+1184,0
+1188,0
+1192,0
+1196,0
+1200,5494
+1204,-856
+1208,-29
+1212,1
+1216,0
+1220,0
+1224,0
+1228,0
+1232,0
+1236,0
+1240,0
+1244,4611
+1248,-909
+1252,-24
+1256,2
+1260,0
+1264,0
+1268,0
+1272,0
+1276,0
+1280,0
+1284,0
+1288,3679
+1292,-952
+1296,-19
+1300,2
+1304,0
+1308,0
+1312,0
+1316,0
+1320,0
+1324,0
+1328,0
+1332,2710
+1336,-986
+1340,-14
+1344,2
+1348,0
+1352,0
+1356,0
+1360,0
+1364,0
+1368,0
+1372,0
+1376,1711
+1380,-1009
+1384,-9
+1388,2
+1392,0
+1396,0
+1400,0
+1404,0
+1408,0
+1412,0
+1416,0
+1420,695
+1424,-1022
+1428,-4
+1432,2
+1436,0
+1440,0
+1444,0
+1448,0
+1452,0
+1456,0
+1460,0
+1464,-328
+1468,-1023
+1472,2
+1476,2
+1480,0
+1484,0
+1488,0
+1492,0
+1496,0
+1500,0
+1504,0
+1508,-1348
+1512,-1015
+1516,7
+1520,2
+1524,0
+1528,0
+1532,0
+1536,0
+1540,0
+1544,0
+1548,0
+1552,-2354
+1556,-995
+1560,12
+1564,2
+1568,0
+1572,0
+1576,0
+1580,0
+1584,0
+1588,0
+1592,0
+1596,-3335
+1600,-965
+1604,17
+1608,2
+1612,0
+1616,0
+1620,0
+1624,0
+1628,0
+1632,0
+1636,0
+1640,-4281
+1644,-925
+1648,22
+1652,2
+1656,0
+1660,0
+1664,0
+1668,0
+1672,0
+1676,0
+1680,0
+1684,-5183
+1688,-876
+1692,27
+1696,2
+1700,0
+1704,0
+1708,0
+1712,0
+1716,0
+1720,0
+1724,0
+1728,-6030
+1732,-817
+1736,32
+1740,1
+1744,0
+1748,0
+1752,0
+1756,0
+1760,0
+1764,0
+1768,0
+1772,-6814
+1776,-750
+1780,36
+1784,1
+1788,0
+1792,0
+1796,0
+1800,0
+1804,0
+1808,0
+1812,0
+1816,-7526
+1820,-674
+1824,39
+1828,1
+1832,0
+1836,0
+1840,0
+1844,0
+1848,0
+1852,0
+1856,0
+1860,-8160
+1864,-592
+1868,43
+1872,1
+1876,0
+1880,0
+1884,0
+1888,0
+1892,0
+1896,0
+1900,0
+1904,-8708
+1908,-503
+1912,46
+1916,1
+1920,0
+1924,0
+1928,0
+1932,0
+1936,0
+1940,0
+1944,0
+1948,-9165
+1952,-410
+1956,48
+1960,1
+1964,0
+1968,0
+1972,0
+1976,0
+1980,0
+1984,0
+1988,0
+1992,-9526
+1996,-312
+2000,50
+2004,1
+2008,0
+2012,0
+2016,0
+2020,0
+2024,0
+2028,0
+2032,0
+2036,-9787
+2040,-210
+2044,51
+2048,0
+2052,0
+2056,0
+2060,0
+2064,0
+2068,0
+2072,0
+2076,0
+2080,-9946
+2084,-107
+2088,52
+2092,0
+2096,0
+2100,0
+2104,0
+2108,0
+2112,0
+2116,0
+2120,0
+2124,-10000
+2128,-2
+2132,52
+2136,0
+2140,0
+2144,0
+2148,0
+2152,0
+2156,0
+2160,0
+2164,0
+2168,-9950
+2172,103
+2176,52
+2180,0
+2184,0
+2188,0
+2192,0
+2196,0
+2200,0
+2204,0
+2208,0
+2212,-9795
+2216,206
+2220,51
+2224,0
+2228,0
+2232,0
+2236,0
+2240,0
+2244,0
+2248,0
+2252,0
+2256,-9538
+2260,308
+2264,50
+2268,-1
+2272,0
+2276,0
+2280,0
+2284,0
+2288,0
+2292,0
+2296,0
+2300,-9181
+2304,406
+2308,48
+2312,-1
+2316,0
+2320,0
+2324,0
+2328,0
+2332,0
+2336,0
+2340,0
+2344,-8727
+2348,500
+2352,46
+2356,-1
+2360,0
+2364,0
+2368,0
+2372,0
+2376,0
+2380,0
+2384,0
+2388,-8183
+2392,589
+2396,43
+2400,-1
+2404,0
+2408,0
+2412,0
+2416,0
+2420,0
+2424,0
+2428,0
+2432,-7552
+2436,671
+2440,40
+2444,-1
+2448,0
+2452,0
+2456,0
+2460,0
+2464,0
+2468,0
+2472,0
+2476,-6843
+2480,747
+2484,36
+2488,-1
+2492,0
+2496,0
+2500,0
+2504,0
+2508,0
+2512,0
+2516,0
+2520,-6061
+2524,814
+2528,32
+2532,-1
+2536,0
+2540,0
+2544,0
+2548,0
+2552,0
+2556,0
+2560,0
+2564,-5217
+2568,874
+2572,27
+2576,-2
+2580,0
+2584,0
+2588,0
+2592,0
+2596,0
+2600,0
+2604,0
+2608,-4317
+2612,924
+2616,23
+2620,-2
+2624,0
+2628,0
+2632,0
+2636,0
+2640,0
+2644,0
+2648,0
+2652,-3372
+2656,964
+2660,18
+2664,-2
+2668,0
+2672,0
+2676,0
+2680,0
+2684,0
+2688,0
+2692,0
+2696,-2392
+2700,994
+2704,13
+2708,-2
+2712,0
+2716,0
+2720,0
+2724,0
+2728,0
+2732,0
+2736,0
+2740,-1387
+2744,1014
+2748,7
+2752,-2
+2756,0
+2760,0
+2764,0
+2768,0
+2772,0
+2776,0
+2780,0
+10000,8414
